+++ jest.config.ts
@@ -1,0 +1,10 @@
+import type { Config } from "@jest/types";
+
+const config: Config.InitialOptions = {
+  preset: "ts-jest",
+  testEnvironment: "node",
+  collectCoverageFrom: ["./src/**/*.ts"],
+  setupFiles: ["./test/polyfills.ts"],
+};
+
+export default config;