+++ test/polyfills.ts
@@ -1,0 +1,5 @@
+import fetch from "node-fetch";
+
+if (!globalThis.fetch) {
+  Object.assign(globalThis, { fetch });
+}